requirements.txt
litestar>=2.0.0
uvicorn>=0.23.2
numpy>=1.25.0,<2
pydantic>=2.4.2
typing-extensions>=4.8.0
onnxruntime>=1.22.1,<2
jsonschema>=4.25.1,<5
requests>=2.32.5,<3
piper-tts>=1.3.0,<2
pyyaml>=6.0,<7
pytest>=8.0.0
cachetools>=5.3.0,<6
loguru>=0.7.3,<0.8
black>=24.0.0,<25
ruff>=0.5.0,<0.6